unsigned vs. size_t code cleanup
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@3576 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/util/identifier.cpp
@@ -52,7 +52,7 @@ void identifiert::parse(const std::string &s)
 {
   std::string component;
 
-  for(unsigned i=0; i<s.size();)
+  for(size_t i=0; i<s.size();)
   {
     for(; i<s.size(); i++)
     {
